+++ sketch-core/src/commonMain/kotlin/com/github/panpf/sketch/util/core_utils.common.kt
@@ -503,3 +503,72 @@ fun ImageOptions?.difference(other: ImageOptions?): String {
     }
     return "Same content"
 }
+
+/**
+ * Compare two versions
+ *
+ * The following version formats are supported:
+ * 1.0, 1.0.0, 1.0.0.1, 1.0.0-snapshot1, 1.0.0-snapshot.1, 1.0.0-snapshot01, 1.0.0-alpha01, 1.0.0-beta01, 1.0.0-rc01
+ *
+ * @see com.github.panpf.sketch.core.common.test.util.CoreUtilsTest.testCompareVersions
+ */
+fun compareVersions(version1: String, version2: String): Int {
+    val (numbers1, suffix1) = version1.split("-", limit = 2)
+        .let { it[0].trim() to it.getOrNull(1)?.trim() }
+    val (numbers2, suffix2) = version2.split("-", limit = 2)
+        .let { it[0].trim() to it.getOrNull(1)?.trim() }
+
+    // Compare numeric parts
+    val numberParts1 = numbers1.split(".")
+    val numberParts2 = numbers2.split(".")
+    val maxLength = maxOf(numberParts1.size, numberParts2.size)
+    var numberCompareResult = 0
+    for (i in 0 until maxLength) {
+        val numberPar1 = numberParts1.getOrNull(i)?.toIntOrNull() ?: 0
+        val numberPart2 = numberParts2.getOrNull(i)?.toIntOrNull() ?: 0
+        if (numberPar1 != numberPart2) {
+            numberCompareResult = numberPar1.compareTo(numberPart2)
+            break
+        }
+    }
+    if (numberCompareResult != 0) {
+        return numberCompareResult
+    }
+
+    // Compare suffix parts
+    val suffixCompareResult: Int
+    if (suffix1 == suffix2) {
+        suffixCompareResult = 0
+    } else if (suffix1 == null) {
+        suffixCompareResult = 1
+    } else if (suffix2 == null) {
+        suffixCompareResult = -1
+    } else {
+        val lowercaseSuffix1 = suffix1.lowercase()
+        val lowercaseSuffix2 = suffix2.lowercase()
+        val suffixTypes = listOf("snapshot", "alpha", "beta", "rc")
+            .mapIndexed { index, s -> index to s }
+        val suffixType1 = suffixTypes.find { lowercaseSuffix1.startsWith(it.second) }
+        val suffixType2 = suffixTypes.find { lowercaseSuffix2.startsWith(it.second) }
+        if (suffixType1 != null && suffixType2 != null) {
+            if (suffixType1 == suffixType2) {
+                val suffix1Number = lowercaseSuffix1
+                    .replace(suffixType1.second, "")
+                    .replace(".", "")
+                    .toIntOrNull() ?: 0
+                val suffix2Number = lowercaseSuffix2.replace(suffixType2.second, "")
+                    .replace(".", "")
+                    .toIntOrNull() ?: 0
+                suffixCompareResult = suffix1Number.compareTo(suffix2Number)
+            } else {
+                suffixCompareResult = suffixType1.first.compareTo(suffixType2.first)
+            }
+        } else if (suffixType1 == null) {
+            suffixCompareResult = 1
+        } else {
+            // suffixType2 == null
+            suffixCompareResult = -1
+        }
+    }
+    return suffixCompareResult
+}