+++ sketch-core/src/commonTest/kotlin/com/github/panpf/sketch/core/common/test/util/CoreUtilsTest.kt
@@ -30,6 +30,7 @@ import com.github.panpf.sketch.util.asOrThrow
 import com.github.panpf.sketch.util.calculateCropBounds
 import com.github.panpf.sketch.util.calculateInsideBounds
 import com.github.panpf.sketch.util.ceilRoundPow2
+import com.github.panpf.sketch.util.compareVersions
 import com.github.panpf.sketch.util.computeScaleMultiplierWithFit
 import com.github.panpf.sketch.util.computeScaleMultiplierWithOneSide
 import com.github.panpf.sketch.util.difference
@@ -858,6 +859,97 @@ class CoreUtilsTest {
                 actual = options.difference(this@apply)
             )
         }
+    }
+
+    @Test
+    fun testCompareVersions() {
+        assertEquals(-1, compareVersions("0.8", "0.8.1"))
+        assertEquals(1, compareVersions("0.8.1", "0.8"))
+        assertEquals(-1, compareVersions("0.8.10", "0.8.10.1"))
+        assertEquals(1, compareVersions("0.8.10.1", "0.8.10"))
+        assertEquals(-1, compareVersions("0.8.15", "0.8.16"))
+        assertEquals(1, compareVersions("0.8.16", "0.8.15"))
+        assertEquals(-1, compareVersions("0.7.99", "0.8.0"))
+        assertEquals(1, compareVersions("0.8.0", "0.7.99"))
+        assertEquals(-1, compareVersions("0.6.99", "0.7.99"))
+        assertEquals(1, compareVersions("0.7.99", "0.6.99"))
+
+        assertEquals(0, compareVersions("1.0.0", "1.0.0"))
+        assertEquals(0, compareVersions("0.8.1", "0.8.1"))
+
+        assertEquals(-1, compareVersions("0.8.0", "0.8.1-SNAPSHOT01"))
+        assertEquals(1, compareVersions("0.8.1-SNAPSHOT01", "0.8.0"))
+        assertEquals(-1, compareVersions("0.8.1-SNAPSHOT01", "0.8.1"))
+        assertEquals(1, compareVersions("0.8.1", "0.8.1-SNAPSHOT01"))
+        assertEquals(-1, compareVersions("0.8.1-SNAPSHOT01", "0.8.2"))
+        assertEquals(1, compareVersions("0.8.2", "0.8.1-SNAPSHOT01"))
+        assertEquals(0, compareVersions("0.8.1-SNAPSHOT01", "0.8.1-SNAPSHOT1"))
+        assertEquals(0, compareVersions("0.8.1-SNAPSHOT09", "0.8.1-SNAPSHOT9"))
+        assertEquals(-1, compareVersions("0.8.1-SNAPSHOT1", "0.8.1-SNAPSHOT2"))
+        assertEquals(-1, compareVersions("0.8.1-SNAPSHOT01", "0.8.1-SNAPSHOT2"))
+        assertEquals(-1, compareVersions("0.8.1-SNAPSHOT01", "0.8.1-SNAPSHOT02"))
+        assertEquals(1, compareVersions("0.8.1-SNAPSHOT2", "0.8.1-SNAPSHOT1"))
+        assertEquals(1, compareVersions("0.8.1-SNAPSHOT2", "0.8.1-SNAPSHOT01"))
+        assertEquals(1, compareVersions("0.8.1-SNAPSHOT02", "0.8.1-SNAPSHOT01"))
+
+        assertEquals(-1, compareVersions("0.8.0", "0.8.1-alpha01"))
+        assertEquals(1, compareVersions("0.8.1-alpha01", "0.8.0"))
+        assertEquals(-1, compareVersions("0.8.1-alpha01", "0.8.1"))
+        assertEquals(1, compareVersions("0.8.1", "0.8.1-alpha01"))
+        assertEquals(-1, compareVersions("0.8.1-alpha01", "0.8.2"))
+        assertEquals(1, compareVersions("0.8.2", "0.8.1-alpha01"))
+        assertEquals(0, compareVersions("0.8.1-alpha01", "0.8.1-alpha1"))
+        assertEquals(0, compareVersions("0.8.1-alpha09", "0.8.1-alpha9"))
+        assertEquals(-1, compareVersions("0.8.1-alpha1", "0.8.1-alpha2"))
+        assertEquals(-1, compareVersions("0.8.1-alpha01", "0.8.1-alpha2"))
+        assertEquals(-1, compareVersions("0.8.1-alpha01", "0.8.1-alpha02"))
+        assertEquals(1, compareVersions("0.8.1-alpha2", "0.8.1-alpha1"))
+        assertEquals(1, compareVersions("0.8.1-alpha2", "0.8.1-alpha01"))
+        assertEquals(1, compareVersions("0.8.1-alpha02", "0.8.1-alpha01"))
+
+        assertEquals(-1, compareVersions("0.8.0", "0.8.1-beta01"))
+        assertEquals(1, compareVersions("0.8.1-beta01", "0.8.0"))
+        assertEquals(-1, compareVersions("0.8.1-beta01", "0.8.1"))
+        assertEquals(1, compareVersions("0.8.1", "0.8.1-beta01"))
+        assertEquals(-1, compareVersions("0.8.1-beta01", "0.8.2"))
+        assertEquals(1, compareVersions("0.8.2", "0.8.1-beta01"))
+        assertEquals(0, compareVersions("0.8.1-beta01", "0.8.1-beta1"))
+        assertEquals(0, compareVersions("0.8.1-beta09", "0.8.1-beta9"))
+        assertEquals(-1, compareVersions("0.8.1-beta1", "0.8.1-beta2"))
+        assertEquals(-1, compareVersions("0.8.1-beta01", "0.8.1-beta2"))
+        assertEquals(-1, compareVersions("0.8.1-beta01", "0.8.1-beta02"))
+        assertEquals(1, compareVersions("0.8.1-beta2", "0.8.1-beta1"))
+        assertEquals(1, compareVersions("0.8.1-beta2", "0.8.1-beta01"))
+        assertEquals(1, compareVersions("0.8.1-beta02", "0.8.1-beta01"))
+
+        assertEquals(-1, compareVersions("0.8.0", "0.8.1-rc01"))
+        assertEquals(1, compareVersions("0.8.1-rc01", "0.8.0"))
+        assertEquals(-1, compareVersions("0.8.1-rc01", "0.8.1"))
+        assertEquals(1, compareVersions("0.8.1", "0.8.1-rc01"))
+        assertEquals(-1, compareVersions("0.8.1-rc01", "0.8.2"))
+        assertEquals(1, compareVersions("0.8.2", "0.8.1-rc01"))
+        assertEquals(0, compareVersions("0.8.1-rc01", "0.8.1-rc1"))
+        assertEquals(0, compareVersions("0.8.1-rc09", "0.8.1-rc9"))
+        assertEquals(-1, compareVersions("0.8.1-rc1", "0.8.1-rc2"))
+        assertEquals(-1, compareVersions("0.8.1-rc01", "0.8.1-rc2"))
+        assertEquals(-1, compareVersions("0.8.1-rc01", "0.8.1-rc02"))
+        assertEquals(1, compareVersions("0.8.1-rc2", "0.8.1-rc1"))
+        assertEquals(1, compareVersions("0.8.1-rc2", "0.8.1-rc01"))
+        assertEquals(1, compareVersions("0.8.1-rc02", "0.8.1-rc01"))
+
+        assertEquals(-1, compareVersions("0.8.0", "0.8.1-SNAPSHOT1"))
+        assertEquals(-1, compareVersions("0.8.1-SNAPSHOT1", "0.8.1-alpha01"))
+        assertEquals(-1, compareVersions("0.8.1-alpha01", "0.8.1-beta1"))
+        assertEquals(-1, compareVersions("0.8.1-beta1", "0.8.1-rc02"))
+        assertEquals(-1, compareVersions("0.8.1-rc02", "0.8.1"))
+        assertEquals(-1, compareVersions("0.8.1", "0.8.2"))
+
+        assertEquals(1, compareVersions("0.8.2", "0.8.1"))
+        assertEquals(1, compareVersions("0.8.1", "0.8.1-rc.02"))
+        assertEquals(1, compareVersions("0.8.1-rc.02", "0.8.1-beta.1"))
+        assertEquals(1, compareVersions("0.8.1-beta.1", "0.8.1-alpha.01"))
+        assertEquals(1, compareVersions("0.8.1-alpha.01", "0.8.1-SNAPSHOT.1"))
+        assertEquals(1, compareVersions("0.8.1-SNAPSHOT.1", "0.8.0"))
     }
 
     private class FormatItem<T>(val number: T, val newScale: Int, val expected: T)